feat: show vision card before plan
